host keys, user backup, update subcommand
This change batches a bunch of things together.

The host keys are now persisted between startups. If there's no existing
host key backup, the startup script will delete the baked-in keys and
re-generate them all before starting anything.

In order to do this efficiently, I changed the structure of the `data`
directory so that it has a top-level directory for keys and a top-level
directory for user data (`passwd`, etc).

Also, added an update subcommand to update the keys/mounts for all
existing users